versao online

+++ app.py
@@ -1,17 +1,89 @@
-from flask import Flask, render_template, request, jsonify
+# app.py
+
+from flask import Flask, render_template, request, jsonify, redirect, session
 import psycopg2
+import os
 
 app = Flask(__name__)
+
+# =========================================
+# SECRET KEY
+# =========================================
+
+app.secret_key = "futebol_dos_amigos"
 
 # =========================================
 # CONEXÃO POSTGRESQL
 # =========================================
 
-import os
-
 DATABASE_URL = os.environ.get("DATABASE_URL")
 
-conn = psycopg2.connect(DATABASE_URL)
+if DATABASE_URL:
+
+    conn = psycopg2.connect(DATABASE_URL)
+
+else:
+
+    conn = psycopg2.connect(
+    host="127.0.0.1",
+    port="5432",
+    dbname="futebol_dos_amigos_db",
+    user="postgres",
+    password="Hm07041997"
+    )
+# =========================================
+# LOGIN
+# =========================================
+
+@app.route("/login")
+def login():
+
+    return render_template("login.html")
+
+# =========================================
+# LOGAR
+# =========================================
+
+@app.route("/logar", methods=["POST"])
+def logar():
+
+    usuario = request.form["usuario"]
+    senha = request.form["senha"]
+
+    cur = conn.cursor()
+
+    cur.execute("""
+        SELECT
+            id,
+            nome,
+            tipo
+        FROM usuarios
+        WHERE usuario = %s
+        AND senha = %s
+    """, (usuario, senha))
+
+    user = cur.fetchone()
+
+    if user:
+
+        session["usuario_id"] = user[0]
+        session["nome"] = user[1]
+        session["tipo"] = user[2]
+
+        return redirect("/")
+
+    return "Usuário inválido"
+
+# =========================================
+# LOGOUT
+# =========================================
+
+@app.route("/logout")
+def logout():
+
+    session.clear()
+
+    return redirect("/login")
 
 # =========================================
 # HOME
@@ -19,7 +91,16 @@ conn = psycopg2.connect(DATABASE_URL)
 
 @app.route("/")
 def home():
-    return render_template("index.html")
+
+    if "usuario_id" not in session:
+
+        return redirect("/login")
+
+    return render_template(
+        "index.html",
+        nome=session["nome"],
+        tipo=session["tipo"]
+    )
 
 # =========================================
 # LISTAR JOGADORES
@@ -27,6 +108,9 @@ def home():
 
 @app.route("/jogadores", methods=["GET"])
 def jogadores():
+
+    if "usuario_id" not in session:
+        return jsonify([])
 
     cur = conn.cursor()
 
@@ -64,6 +148,9 @@ def jogadores():
 @app.route("/jogadores", methods=["POST"])
 def add_jogador():
 
+    if session["tipo"] != "admin":
+        return jsonify({"erro": "Sem permissão"})
+
     data = request.json
 
     nome = data["nome"].upper()
@@ -74,13 +161,14 @@ def add_jogador():
 
     cur = conn.cursor()
 
-    cur.execute(
-        """
-        INSERT INTO jogadores(nome, nota, posicao)
-        VALUES(%s, %s, %s)
-        """,
-        (nome, nota, posicao)
-    )
+    cur.execute("""
+        INSERT INTO jogadores (
+            nome,
+            nota,
+            posicao
+        )
+        VALUES (%s, %s, %s)
+    """, (nome, nota, posicao))
 
     conn.commit()
 
@@ -95,18 +183,18 @@ def add_jogador():
 @app.route("/gol", methods=["POST"])
 def gol():
 
+    if session["tipo"] != "admin":
+        return jsonify({"erro": "Sem permissão"})
+
     data = request.json
 
     cur = conn.cursor()
 
-    cur.execute(
-        """
+    cur.execute("""
         UPDATE jogadores
         SET gols = gols + 1
         WHERE id = %s
-        """,
-        (data["id"],)
-    )
+    """, (data["id"],))
 
     conn.commit()
 
@@ -121,21 +209,21 @@ def gol():
 @app.route("/remover-gol", methods=["POST"])
 def remover_gol():
 
+    if session["tipo"] != "admin":
+        return jsonify({"erro": "Sem permissão"})
+
     data = request.json
 
     cur = conn.cursor()
 
-    cur.execute(
-        """
+    cur.execute("""
         UPDATE jogadores
         SET gols = CASE
             WHEN gols > 0 THEN gols - 1
             ELSE 0
         END
         WHERE id = %s
-        """,
-        (data["id"],)
-    )
+    """, (data["id"],))
 
     conn.commit()
 
@@ -150,17 +238,17 @@ def remover_gol():
 @app.route("/excluir-jogador", methods=["POST"])
 def excluir_jogador():
 
+    if session["tipo"] != "admin":
+        return jsonify({"erro": "Sem permissão"})
+
     data = request.json
 
     cur = conn.cursor()
 
-    cur.execute(
-        """
+    cur.execute("""
         DELETE FROM jogadores
         WHERE id = %s
-        """,
-        (data["id"],)
-    )
+    """, (data["id"],))
 
     conn.commit()
